backend/model/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pyyaml
pydantic
psycopg2-binary
